Refine CLI and HTTP domain boundaries
Move list-scoped book reads into the books service and repository so
list responses share normal book hydration behavior.

- Share CLI database setup and add presence-aware optional flags
- Enable SQLite foreign keys on every pooled connection
- Split CLI commands, HTTP handlers, and endpoint tests into focused
  domain files
- Preserve injected output writers and handle ignored I/O errors
