+++ internal/lists/sqlite_repository_test.go
@@ -183,55 +183,3 @@ func TestSQLiteRepositoryAddBookToListReturnsErrBookNotFound(t *testing.T) {
 		t.Fatalf("expected ErrBookNotFound, got %v", err)
 	}
 }
-
-// ── ListBooks ─────────────────────────────────────────────────────────────────
-
-func TestSQLiteRepositoryListBooksReturnsBooksInList(t *testing.T) {
-	ctx := context.Background()
-	db := testsupport.OpenMigratedDB(t)
-	repository := lists.NewSQLiteRepository(db)
-
-	listID := testsupport.InsertListRow(t, db, "Want to Buy")
-	bookID1 := testsupport.InsertBookRow(t, db, "Dune", nil)
-	bookID2 := testsupport.InsertBookRow(t, db, "Foundation", nil)
-
-	if err := repository.AddBookToList(ctx, listID, bookID1); err != nil {
-		t.Fatal(err)
-	}
-	if err := repository.AddBookToList(ctx, listID, bookID2); err != nil {
-		t.Fatal(err)
-	}
-
-	bookList, err := repository.ListBooks(ctx, listID)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(bookList) != 2 {
-		t.Fatalf("expected 2 books, got %d", len(bookList))
-	}
-	if bookList[0].Title != "Dune" {
-		t.Fatalf("expected Dune, got %q", bookList[0].Title)
-	}
-	if bookList[1].Title != "Foundation" {
-		t.Fatalf("expected Foundation, got %q", bookList[1].Title)
-	}
-}
-
-func TestSQLiteRepositoryListBooksReturnsEmptySliceForEmptyList(t *testing.T) {
-	ctx := context.Background()
-	db := testsupport.OpenMigratedDB(t)
-	repository := lists.NewSQLiteRepository(db)
-
-	listID := testsupport.InsertListRow(t, db, "Want to Buy")
-
-	bookList, err := repository.ListBooks(ctx, listID)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if bookList == nil {
-		t.Fatal("expected non-nil slice")
-	}
-	if len(bookList) != 0 {
-		t.Fatalf("expected empty slice, got %d books", len(bookList))
-	}
-}
